+++ load.py
@@ -4,6 +4,7 @@ from semantic_version import Version #type: ignore
 import tkinter as tk
 
 import myNotebook as nb  #type: ignore
+import edmc_data # type: ignore
 
 from Router.constants import GH_PROJECT, NAME, TITLE, errs, CarrierStates
 from utils.debug import Debug, catch_exceptions
@@ -16,6 +17,7 @@ from Router.csv import CSV
 from Router.ui import UI
 from Router.overlay import Overlay
 from Router.hotkeys import Hotkeys
+from Router.prefs import Prefs
 
 def plugin_start3(plugin_dir: str) -> str:
     Debug(plugin_dir)
@@ -35,19 +37,21 @@ def plugin_start3(plugin_dir: str) -> str:
 
     return NAME
 
-
+@catch_exceptions
 def plugin_start(plugin_dir: str) -> None:
     """EDMC calls this function when running in Python 2 mode."""
     raise EnvironmentError(errs["required_version"])
 
 
+@catch_exceptions
 def plugin_stop() -> None:
     Context.router.save()
+    Context.overlay.stop_countdowns()
     if Context.updater.install_update:
         Context.updater.install()
 
-
 def plugin_app(parent:tk.Widget) -> tk.Frame:
+    Context.prefs = Prefs()
     Context.csv = CSV()
     Context.router = Router()
     Context.ui = UI(parent)
@@ -56,22 +60,22 @@ def plugin_app(parent:tk.Widget) -> tk.Frame:
     if Context.route.route != []:
         Context.overlay.show_frame('Default')
 
-    parent.after(1000, Context.overlay.update_jump_overlay)
+    parent.after(1000, Context.overlay.update_overlays)
     return Context.ui.frame
 
-
+@catch_exceptions
 def journal_entry(cmdr:str, is_beta:bool, system:str, station:str, entry:dict, state:dict) -> None:
     if Context.router == None: return
 
     match entry['event']:
         case 'Startup':
             Context.router.carrier_state = CarrierStates.Idle
-            if Context.route.route != []:
+            if Context.route.route != [] and not Context.route.fleetcarrier:
                 Context.route.update_route(0, system)
                 Context.route.jumps = []
         case 'FSDJump' | 'Location' | 'SupercruiseExit' if entry.get('StarSystem', system) != Context.router.system:
             Context.router.jumped(system, entry)
-        case 'CarrierJumpRequest' | 'CarrierLocation' | 'CarrierJumpCancelled':
+        case 'CarrierJumpRequest' | 'CarrierLocation' | 'CarrierJumpCancelled' | 'CarrierStats':
             Context.router.carrier_event(entry)
         case 'Loadout':
             Context.router.set_ship(entry)
@@ -85,7 +89,7 @@ def journal_entry(cmdr:str, is_beta:bool, system:str, station:str, entry:dict, s
                     case "next":
                         Context.router.update_route(1)
                     case _:
-                        copy_to_clipboard(Context.ui.parent, Context.route.next_stop())
+                        copy_to_clipboard(Context.ui.parent, Context.route.next_system())
         case 'Refueling': # Read fuel from Status.json
             Context.router.fuel_event(state)
         case 'Shutdown':
@@ -93,20 +97,24 @@ def journal_entry(cmdr:str, is_beta:bool, system:str, station:str, entry:dict, s
             Context.router.save()
 
     Context.router.system = system
-    Context.router.cargo = sum(state.get('Cargo', {}).values())
-    Context.ui.update_cargo(Context.router.cargo)
+    cargo:int = sum(state.get('Cargo', {}).values())
+    if cargo != Context.router.cargo:
+        Context.router.cargo = cargo
+        Context.ui.update_cargo(cargo)
 
 
+@catch_exceptions
 def dashboard_entry(cmdr:str, is_beta:bool, entry:dict) -> None:
-    if Context.router:
-        Context.router.dashboard_entry(cmdr, is_beta, entry)
+    if Context.ui.parent and Context.route.jumps_remaining() and entry.get("GuiFocus") == edmc_data.GuiFocusGalaxyMap:
+        copy_to_clipboard(Context.ui.parent, Context.route.next_system())
+
     if Context.overlay:
         Context.overlay.dashboard_entry(cmdr, is_beta, entry)
 
-
+@catch_exceptions
 def plugin_prefs(parent:tk.Frame, cmdr: str, is_beta: bool) -> nb.Frame:
-    return Context.ui.prefs_frame(parent)
+    return Context.prefs.prefs_frame(parent)
 
-
+@catch_exceptions
 def prefs_changed(cmdr: str, is_beta: bool) -> None:
-    Context.ui.save_prefs()
+    Context.prefs.save_prefs()